カメラ要素とUIの接続
